pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "python-docx>=1.1.0",
    "flask>=3.1.1",
    "gdown>=5.2.0",
    "pandas>=2.0.0",
    "textract>=1.6.5",
    "PyMuPDF>=1.23.0",
    "python-magic>=0.4.27",
    "gunicorn>=21.2.0"
]
